extractors/youtube: Separate into a package
