+++ extractors/youtube/signature.go
@@ -1,4 +1,4 @@
-package extractors
+package youtube
 
 import (
 	"fmt"
